% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altGenomenclature.R
\name{altGenomenclature}
\alias{altGenomenclature}
\title{Obtain alternate genome nomenclature}
\usage{
altGenomenclature(species, genome)
}
\arguments{
\item{species}{String of species name/alias.}

\item{genome}{String of the genome version used. Nomenclature from UCSC Genome Browser and Genome Reference Consortium are both acceptable inputs.}
}
\value{
List of both genome identifiers. The first element in the list is the GRC nomenclature, and the second element is the UCSC nomenclature.
}
\description{
\code{altGenomenclature} returns a list containing the genome identifiers of the genome version provided.
}
\examples{
altGenomenclature("mouse", "mm10")
}
