% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpG_barplot.R
\name{CpG_barplot}
\alias{CpG_barplot}
\title{Get bar plot visualizing amounts of CpGs within a genomic region}
\usage{
CpG_barplot(chr, start, end)
}
\arguments{
\item{chr}{integer number chosen chromosome}

\item{start}{integer defining the start position of the analysed genomic region}

\item{end}{integer defining the end position of the analysed genomic region}
}
\value{
bar plot
}
\description{
Get bar plot visualizing amounts of CpGs within a genomic region
}
