% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_meta.R
\docType{data}
\name{ex_meta}
\alias{ex_meta}
\title{Meta data for example sparse matrices}
\format{
A data frame containing 493 rows and 2 columns
\describe{
\item{cell_id_dna}{cell id}
\item{ptime}{pseudo time 1}
\item{ptime2}{pseudo time 2}
\item{pt_avg}{average pseudo time}}
}
\usage{
data(ex_meta)
}
\description{
Data frame with meta data for example sparse matrix ex_spM and ex_spMacc, containing cell ids and pseudotimes.
}
\references{
?
}
\keyword{datasets}
