% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_spM.R
\docType{data}
\name{ex_spM}
\alias{ex_spM}
\title{Example dgTMatrix of sample CpG methylation data}
\format{
dgTMatrix with dimensions 70001 x 833
}
\source{
Sampling of a sparse matrix for chromosome 8 \code{spM[8620000:8690000, 1:833]} from scNMT-seq data of \verb{Mus musculus} received from \href{https://doi.org/10.1101/2022.07.13.499860}{(Kremer et al., 2022)}.
}
\usage{
data(ex_spM)
}
\description{
Sparse Matrix of format \code{dgTMatrix} containing sample CpG methylation data of astrocytes and the NSC lineage obtained via scNMT-seq.
}
\references{
Kremer, L. P., Cerrizuela, S., Shukairi, M. E. A., Ellinger, T., Straub, J., Dehler, S., Korkmaz, A., Weichenhan, D., Plass, C., Anders, S., & Martin-Villalba, A. (2022). \if{html}{\out{<br>}}
Single-cell triple-omics uncovers DNA methylation as key feature of stemness in the healthy and ischemic adult brain. \if{html}{\out{<br>}}
bioRxiv, 2022.2007.2013.499860 \url{https://doi.org/10.1101/2022.07.13.499860}.
}
\keyword{datasets}
