% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_spMacc.R
\docType{data}
\name{ex_spMacc}
\alias{ex_spMacc}
\title{dgTMatrix containing chromatin accessibility data of scNMT-seq}
\format{
dgTMatrix with dimensions 70001 x 797
}
\usage{
data(ex_spMacc)
}
\description{
dgTMatrix containing GpC methylation data of astrocytes and the NSC lineage.
}
\examples{
\dontrun{map_methyl( ex_spMacc, ex_meta, ex_headeracc, 8000, 64000)}
}
\references{
?
}
\keyword{datasets}
