% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genemodel.R
\name{get_genemodel}
\alias{get_genemodel}
\title{Obtain Ensembl genemodel}
\usage{
get_genemodel(genepath, chr, startpos, endpos)
}
\arguments{
\item{genepath}{Genome of the chosen species retrieved from the ensemble FTP site \url{https://ftp.ensembl.org/pub/}.
Provide either a single character string containing the path or URL of a GFF file, or a GFF-derived data frame created with readGFF() or returned by get_regfeat().}

\item{chr}{Integer number of chromosome.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}
}
\value{
Data frame containing various columns, depending on the chosen genome version and release.
Possible column names include \code{seqname}, \code{start}, \code{end}, \code{width}, \code{strand}, \code{source}, \code{type} , \code{gene_id} and \code{strand_boolean}.
}
\description{
\code{get_genemodel} downloads the gene annotations between \code{startpos} and \code{endpos} of the chromosome \code{chr} from the chosen species and removes transcript annotations.
Gene annotation includes genes, exons, CDS, start codons, stop codons, five prime utrs, three prime utrs and Selenocysteines.
For this, copy the link or download the GTF file from the Ensembl FTP site \url{https://ftp.ensembl.org/pub/}.
}
\examples{
genes_url <- "https://ftp.ensembl.org/pub/release-110/gtf/mus_musculus/Mus_musculus.GRCm39.110.gtf.gz"
get_genemodel(genes_url, 8,  8628165, 8684055)

}
\seealso{
\code{\link[=plot_genemodel]{plot_genemodel()}} to plot data frame as an annotated arrow plot.
}
