% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regfeat.R
\name{get_regfeat}
\alias{get_regfeat}
\title{Obtain Ensembl regulatory features}
\usage{
get_regfeat(regpath, chr, startpos, endpos)
}
\arguments{
\item{regpath}{Regulatory features of the chosen species retrieved from the ensemble FTP site \url{https://ftp.ensembl.org/pub/}.
Provide either a single character string containing the path or URL of a GFF file, or a GFF-derived data frame created with readGFF() or returned by get_regfeat().}

\item{chr}{Integer number of chromosome.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}
}
\value{
A data frame containing the regulatory features and the following 14 columns:
\code{seqnames}, \code{start}, \code{end}, \code{width}, \code{strand}, \code{source}, \code{type}, \code{score}, \code{phase}, \code{ID}, \code{bound_end}, \code{bound_start}, \code{description}, and \code{feature_type}.
If the input is a data frame created with \code{readGFF()}, the \code{width} column is not included.
}
\description{
\code{get_regfeat} downloads regulatory features found between \code{startpos} and \code{endpos} of the chromosome \code{chr} from the chosen Ensembl dataset.
For this, copy the link or download the GTF file from the Ensembl FTP site \url{https://ftp.ensembl.org/pub/}.
Regulatory features include promoters, enhancers, open chromatin regions, transcription factor binding and CTF binding sites.
More information here: \url{https://www.ensembl.org/info/genome/funcgen/data/regulatory-features.html}.
}
\examples{
reg_url <- "https://ftp.ensembl.org/pub/release-110/regulation/mus_musculus/mus_musculus.GRCm39.Regulatory_Build.regulatory_features.20221007.gff.gz"
get_regfeat(reg_url, 8, 8628165, 8684055)


}
\seealso{
\code{\link[=plot_regfeat]{plot_regfeat()}} to plot data fame as an annotated segment plot.
}
