% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exdata.R
\name{load_exdata}
\alias{load_exdata}
\title{Download example data}
\source{
Original files received from \href{https://doi.org/10.1101/2022.07.13.499860}{(Kremer et al., 2022)}. The sparse matrices were simplified by removing the values outside the chosen region.
}
\usage{
load_exdata()
}
\value{
List of directories, in which the temporary files are saved.
}
\description{
\code{load_exdata} downloads and saves example data as temporary files. The list of directories returned contain the files in order:
\describe{
\item{8_trunc.npz}{Numpy sparse matrix generated by \code{methSCAn} python package with dimensions 129300644x833. Contains CpG methylation of chromosome 8 from a murine adult NSC lineage. Information was reduced to the genomic region 8628100-8684100.}
\item{8_GpC_trunc.npz}{Numpy sparse matrix generated by \code{methSCAn} python package with dimensions 1129301117x797. Contains GpC methylation of chromosome 8 from a murine adult NSC lineage. Information was reduced to the genomic region 8628100-8684100.}
\item{cell-annotation.tsv}{Tsv file containing cell ids and pseudotime of the cells used for the scNMT-sequencing to receive the original Numpy sparse matrices.}
}
}
\examples{
load_exdata()
}
\references{
Kremer, L. P., Cerrizuela, S., Shukairi, M. E. A., Ellinger, T., Straub, J., Dehler, S., Korkmaz, A., Weichenhan, D., Plass, C., Anders, S., & Martin-Villalba, A. (2022). \if{html}{\out{<br>}}
Single-cell triple-omics uncovers DNA methylation as key feature of stemness in the healthy and ischemic adult brain. \if{html}{\out{<br>}}
bioRxiv, 2022.2007.2013.499860 \url{https://doi.org/10.1101/2022.07.13.499860}.
}
