% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methyl.R
\name{map_methyl}
\alias{map_methyl}
\title{Map sparse matrix to pseudotime}
\usage{
map_methyl(spM, meta, header, startpos, endpos)
}
\arguments{
\item{spM}{TsparseMatrix containing sparse methylation data (either CpG or GpC) from scNMT-seq with j = indices, p = indptr and x = data. Can be generated from npz-files using \code{npz.to.spM}.}

\item{meta}{Data frame of meta data containing cell IDs ("cell_id_dna") and pseudotime ("ptime").}

\item{header}{Data frame with a column \code{cell_id} mapping the sparse matrix column indices to cell IDs. This must match the format used in \code{meta$cell_id_dna}.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}
}
\value{
Data frame with three columns, containing information of spM for the selected genomic region, sorted by ptime.
\describe{
\item{pos}{Genomic position (row index of the matrix).}
\item{pt}{Pseudotime value for the corresponding cell.}
\item{methyl}{Methylation status: -1 (unmethylated), 1 (methylated).}
}
}
\description{
\code{map_methyl} maps the sparse methylation matrix to its meta data. It uses the \code{header} to assign column names (cell IDs) to the sparse Matrix, then sorts and maps the matrix columns to pseudotimes based on \code{meta}.
}
\examples{
map_methyl(ex_spM, ex_meta, ex_header, 8000, 64000)
}
