% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npz.to.spMatrix.R
\name{npz.to.spMatrix}
\alias{npz.to.spMatrix}
\title{Transform npz file to sparse Matrix}
\usage{
npz.to.spMatrix(npzPath)
}
\arguments{
\item{npzPath}{pathway of npz file}
}
\description{
Transform npz file to sparse Matrix
}
