% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genemodel.R
\name{plot_genemodel}
\alias{plot_genemodel}
\title{Plot Ensembl genemodel}
\usage{
plot_genemodel(genepath, chr, startpos, endpos)
}
\arguments{
\item{genepath}{Genome of the chosen species retrieved from the ensemble FTP site \url{https://ftp.ensembl.org/pub/}.
Provide either a single character string containing the path or URL of a GFF file, or a GFF-derived data frame created with readGFF() or returned by get_regfeat().}

\item{chr}{Integer number of chromosome.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}
}
\value{
Arrow plot of genes including colored annotations of the genomic features. Arrows directed to the left are minus strand genes and arrows directed to the right are plus strand genes.
}
\description{
\code{plot_genemodel} downloads the gene annotations between \code{startpos} and \code{endpos} of the chromosome \code{chr} from the chosen species and plots the data as a arrow plot.
Gene annotation includes genes, exons, CDS, start codons, stop codons, five prime utrs, three prime utrs and Selenocysteines.
For this, copy the link or download the GTF file from the Ensembl FTP site \url{https://ftp.ensembl.org/pub/}.
}
\examples{
genes_url <- "https://ftp.ensembl.org/pub/release-110/gtf/mus_musculus/Mus_musculus.GRCm39.110.gtf.gz"
plot_genemodel(genes_url, 8, 8628165, 8684055)

}
\seealso{
\code{\link[=get_genemodel]{get_genemodel()}} to receive the data frame used for plotting.
}
