% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methyl.R
\name{plot_methyl}
\alias{plot_methyl}
\title{Plot mapped methylation data}
\usage{
plot_methyl(m, mappedpt, startpos)
}
\arguments{
\item{m}{Matrix received from \code{smooth2d}. Containes smoothed values and pseudotime of 2 dimensional scNMT-seq data.}

\item{mappedpt}{Data frame containing pseudotimes received from \code{map_methyl}.}

\item{startpos}{Integer defining the start position of the analysed genomic region.}
}
\value{
Tile plot visualizing methylation pattern of a genomic region at various differentiation stages.
}
\description{
\code{plot_methyl} plots a matrix containing the smoothed values of scNMT-seq as a tile plot to visualizing methylation changes during differentiation.
}
\examples{
\dontrun{plot_methyl(m, mappedpt, 8628165)}
}
