% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regfeat.R
\name{plot_regfeat}
\alias{plot_regfeat}
\title{Plot Ensembl regulatory features}
\usage{
plot_regfeat(regpath, chr, startpos, endpos, start_VMR = 0, end_VMR = 0)
}
\arguments{
\item{regpath}{Regulatory features of the chosen species retrieved from the ensemble FTP site \url{https://ftp.ensembl.org/pub/}.
Provide either a single character string containing the path or URL of a GFF file, or a GFF-derived data frame created with readGFF() or returned by get_regfeat().}

\item{chr}{Integer number of chromosome.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}

\item{start_VMR, end_VMR}{Integers defining the start and end position of the variable methylated region. Default is 0.}
}
\value{
Segment plot of regulatory features with colored annotation of a variable methylated region, if provided.
}
\description{
\code{plot_regfeat} downloads regulatory features found between \code{startpos} and \code{endpos} of the chromosome \code{chr} from the chosen ensemble dataset and plots the data as a segment plot.
For this, copy the link or download the GTF file from the Ensembl FTP site \url{https://ftp.ensembl.org/pub/}.
Regulatory features include promoters, enhancers, open chromatin regions, transcription factor binding and CTF binding sites.
More information here: \url{https://www.ensembl.org/info/genome/funcgen/data/regulatory-features.html}.
}
\examples{
reg_url <- "https://ftp.ensembl.org/pub/release-110/regulation/mus_musculus/mus_musculus.GRCm39.Regulatory_Build.regulatory_features.20221007.gff.gz"
plot_regfeat(reg_url, 8, 8628165, 8684055)
}
\seealso{
\code{\link[=get_regfeat]{get_regfeat()}} to receive the data frame used for plotting.
}
