% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regfeatures.R
\name{plot_regfeatures}
\alias{plot_regfeatures}
\title{Plot regulatory features of genomic region}
\usage{
plot_regfeatures(gffpath, chr, startpos, endpos, start_VR = 0, end_VR = 0)
}
\arguments{
\item{gffpath}{Location of the gff file to be read. Can be a single string of the file path or the URL or can be a connection.}

\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}

\item{start_VR}{integer defining the start position of the variable region}

\item{end_VR}{integer defining the end position of the variable region}
}
\value{
segment plot
}
\description{
Plot regulatory features of genomic region
}
\examples{
plot_regfeatures("https://ftp.ensembl.org/pub/release-110/regulation/mus_musculus/mus_musculus.GRCm39.Regulatory_Build.regulatory_features.20221007.gff.gz", 8, 8628165, 8684055, 8653165, 8659055)
}
