% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptime_order.R
\name{ptime_order}
\alias{ptime_order}
\title{Create data frame from sparse Matrix}
\usage{
ptime_order(spM, meta, header, start, end)
}
\arguments{
\item{spM}{sparse Matrix}

\item{meta}{meta data}

\item{header}{column header for spM}

\item{start}{integer defining the start position of the analysed genomic region}

\item{end}{integer defining the end position of the analysed genomic region}
}
\value{
data frame containing information of spM for the selected genomic region, sorted by ptime. Here i is the row index giving information of the genomic position , ptimes contains the sorted pseudo times of the cells and data contains either value -1 or 1 specifying the methylation status (unmethylated, methylated).
}
\description{
Create data frame from sparse Matrix
}
