% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_header.R
\docType{data}
\name{ex_header}
\alias{ex_header}
\title{Cell ids for ex_spM}
\format{
A data frame containing 833 rows and 1 column
\describe{
\item{cell_id}{cell identifier} }
}
\usage{
data(ex_header)
}
\description{
Data frame with cell ids for the used sparse Matrix ex_spM.
}
\examples{
\dontrun{map_methyl( ex_spM, ex_meta, ex_header, 8000, 64000)}
}
\references{
?
}
\keyword{datasets}
