% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_headeracc.R
\docType{data}
\name{ex_headeracc}
\alias{ex_headeracc}
\title{Cell ids for ex_spMacc}
\format{
A data frame containing 797 rows and 1 column
\describe{
\item{cell_id}{cell id} }
}
\usage{
data(ex_headeracc)
}
\description{
Data frame with cell ids for the used sparse Matrix ex_spMacc
}
\examples{
\donttest{map_methyl( ex_spMacc, ex_meta, ex_headeracc, 8000, 64000)}
}
\references{
?
}
\keyword{datasets}
