% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_meta.R
\docType{data}
\name{ex_meta}
\alias{ex_meta}
\title{Meta data of example sparse matrix}
\format{
A data frame containing 493 rows and 2 columns
\describe{
\item{cell_id_dna}{cell id}
\item{ptime}{pseudotime 1}
\item{ptime2}{pseudotime 2}
\item{pt_avg}{average pseudotime}}
}
\source{
Data frame containing cell annotations from \href{https://doi.org/10.1038/s41586-024-07898-9}{(Kremer, Cerrizuela, et al., 2024)}. Additionally, average values of the pseudotimes were calculated.
}
\usage{
data(ex_meta)
}
\description{
Data frame of meta data for example sparse matrix ex_spM, containing cell ids and pseudotimes.
}
\references{
Kremer, L. P. M., Cerrizuela, S., El-Sammak, H., Al Shukairi, M. E., Ellinger, T., Straub, J., Korkmaz, A., Volk, K., Brunken, J., Kleber, S., Anders, S., & Martin-Villalba, A. (2024). DNA methylation controls stemness of astrocytes in health and ischaemia. Nature, 634(8033), 415-423. https://doi.org/10.1038/s41586-024-07898-9
}
\keyword{datasets}
