% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_meta.R
\docType{data}
\name{ex_meta}
\alias{ex_meta}
\title{Meta data for example sparse Matrices}
\format{
A data frame containing 493 rows and 2 columns
\describe{
\item{cell_id_dna}{cell id}
\item{ptime}{pseudo time} }
}
\usage{
data(ex_meta)
}
\description{
Data frame with meta data for example sparse Matrix ex_spM and ex_spMacc
}
\examples{
\donttest{map_methyl( ex_spM, ex_meta, ex_header, 8000, 64000)}
}
\references{
?
}
\keyword{datasets}
