% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_spM.R
\docType{data}
\name{ex_spM}
\alias{ex_spM}
\title{dgTMatrix containing CpG methylation data of scNMT-seq}
\format{
dgTMatrix with dimensions 70001 x 833
}
\usage{
data(ex_spM)
}
\description{
dgTMatrix containing CpG methylation data of astrocytes and the NSC lineage.
}
\references{
?
}
\keyword{datasets}
