% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_kernel_2d.R
\name{gauss_kernel_2d}
\alias{gauss_kernel_2d}
\title{Gaussian Filter applied to 2 dimensional methylation data}
\usage{
gauss_kernel_2d(
  mappedpt,
  hx,
  ht,
  delx = (hx/10),
  delt = (ht/10),
  xrange = range(mappedpt$pos),
  trange = range(mappedpt$pt)
)
}
\arguments{
\item{mappedpt}{data frame received from function \emph{ptime_order} containing relative genomic position (pos), pseudo time of the cells (pt) and methylation status (methyl) -1 or 1 (un-/methylated).}

\item{hx}{width of kernel}

\item{ht}{height of kernel}

\item{delx}{width of grids}

\item{delt}{height of grids}

\item{xrange}{vector containing minimum and maximum of the genomic region}

\item{trange}{vector containing minimum and maximum of the pseudo time}
}
\value{
matrix ...
}
\description{
Gaussian Filter applied to 2 dimensional methylation data
}
\examples{
\dontrun{gauss_kernel_2d(mappedpt, 400, 0.8, 40, 0.08,  xrange = range(mappedpt$pos), trange=range(mappedpt$pt)}
}
