% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpgs.R
\name{get_cpgs}
\alias{get_cpgs}
\title{Get Positions of CpGs in a genomic region}
\usage{
get_cpgs(chr, startpos, endpos)
}
\arguments{
\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}
}
\value{
integer list
}
\description{
Get Positions of CpGs in a genomic region
}
\examples{
get_cpgs(8, 8628165, 8684055)
}
