% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpgs.R
\name{get_cpgs}
\alias{get_cpgs}
\title{Obtain positions of CGs}
\usage{
get_cpgs(species, chr, startpos, endpos)
}
\arguments{
\item{species}{String of species name/alias.}

\item{chr}{Integer number of chromosome.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}
}
\value{
Integer list containing genomic positions of each CG site found within the given region.
}
\description{
\code{get_cpgs} downloads the sequence of the region of interest for a chosen species from \url{https://rest.ensembl.org} and extracts the positions of all CG sites found.
}
\examples{
get_cpgs("mouse", 8, 8628165, 8684055)
}
\seealso{
\code{\link[=plot_cpgs]{plot_cpgs()}} to plot integer list as a bar plot.
}
