% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CpGs.R
\name{get_CpGs}
\alias{get_CpGs}
\title{Get Positions of CpGs in a genomic region}
\usage{
get_CpGs(chr, start, end)
}
\arguments{
\item{chr}{integer number of chromosome}

\item{start}{integer defining the start position of the analysed genomic region}

\item{end}{integer defining the end position of the analysed genomic region}
}
\value{
integer list
}
\description{
Get Positions of CpGs in a genomic region
}
