% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genemodel.R
\name{get_genemodel}
\alias{get_genemodel}
\title{Get gene model of genomic region}
\usage{
get_genemodel(gtf, chr, startpos, endpos)
}
\arguments{
\item{gtf}{pathway to gtf file downloaded from...}

\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}
}
\value{
data frame
}
\description{
Get gene model of genomic region
}
