% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regfeatures.R
\name{get_regfeatures}
\alias{get_regfeatures}
\title{Get regulatory features of genomic region}
\usage{
get_regfeatures(regfeatures, chr, start_pos, end_pos)
}
\arguments{
\item{regfeatures}{pathway of gtf file downloaded from ensemble containing ...}
}
\value{
data frame ...
}
\description{
Get regulatory features of genomic region
}
