% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regfeatures.R
\name{get_regfeatures}
\alias{get_regfeatures}
\title{Get regulatory features of genomic region}
\usage{
get_regfeatures(regfeatures, chr, startpos, endpos)
}
\arguments{
\item{regfeatures}{pathway of gtf file downloaded from ensemble containing ...}

\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}
}
\value{
data frame ...
}
\description{
Get regulatory features of genomic region
}
