% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regulation.R
\name{get_regulation}
\alias{get_regulation}
\title{Obtain ensemble regulatory features}
\usage{
get_regulation(featurepath, chr, startpos, endpos)
}
\arguments{
\item{featurepath}{Location of the gff file to be read. Can be a single string of the file directory, the URL or can be a connection. Contains regulatory features of the chosen species retrieved from the ensemble FTP site https://ftp.ensembl.org/pub/.}

\item{chr}{Integer number of chromosome.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}
}
\value{
Data frame containing 14 columns (seqnames, start, end, width, strand, source, type, score, phase, ID, bound_end, bound_start, description, feature_type).
}
\description{
\code{get_regulation} downloads regulatory features found between \code{startpos}and \code{endpos} of the chromosome \code{chr} from the ensemble dataset chosen.
For this, copy the link or download the GTF file of the ensemble FTP site https://ftp.ensembl.org/pub/.
Regulatory features include promoters, enhancers, open chromatin regions, transcription factor binding and CTF binding sites.
More information here: https://www.ensembl.org/info/genome/funcgen/data/regulatory-features.html.
}
\examples{
get_regulation("https://ftp.ensembl.org/pub/release-110/regulation/mus_musculus/mus_musculus.GRCm39.Regulatory_Build.regulatory_features.20221007.gff.gz", 8, 8628165, 8684055)
}
\seealso{
\code{\link[=plot_genemodel]{plot_genemodel()}} to plot data fame as an annotated segment plot.
}
