% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_exdata.R
\name{load_exdata}
\alias{load_exdata}
\title{Download example data}
\usage{
load_exdata()
}
\value{
List of directories, in which the temporary files are saved.
}
\description{
\code{load_exdata} downloads and saves example data as temporary files.
\enumerate{
\item Npz file containing CpG methylation of chromosome 8 from scNMT analysis.
\item Npz file containing GpC methylation of chromosome 8 from scNMT analysis.
\item Tsv file containing cell ids and pseudotime.
}
}
\examples{
load_exdata()
}
\references{
?
}
