% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methyl.R
\name{map_methyl}
\alias{map_methyl}
\title{Map sparse matrix to pseudotime}
\usage{
map_methyl(spM, meta, header, startpos, endpos)
}
\arguments{
\item{spM}{Tsparse matrix with j = indices, p = indptr and x = data. Can be received using \code{npz.to.spM}. Contains either CpG or GpC methylation data from scNMT-seq.}

\item{meta}{Data frame of meta data containing cell IDs ("cell_id_dna") and pseudotime ("ptime").}

\item{header}{Data frame containing cell IDs for spM. Cell IDs have to have same format as in meta.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}
}
\value{
Data frame containing information of spM for the selected genomic region, sorted by ptime. Here, \code{pos} is the row index indicating the genomic position, \code{pt} contains the sorted pseudotimes and \code{methyl} contains either -1 or 1, specifying the methylation status (un-/methylated).
}
\description{
\code{map_methyl} maps the sparse Matrix containing the methylation data to its meta data. For this, the cell ids are retrieved from \code{cellids} and put as column names of \code{spM}. \code{meta} is used to sort the cell ids according to their pseudotimes and map the pseudotimes to the methylation data of the genomic region of interest.
}
\examples{
map_methyl(ex_spM, ex_meta, ex_header, 8000, 64000)
}
