% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npz.to.spM.R
\name{npz.to.spM}
\alias{npz.to.spM}
\title{Convert npz file to sparse Matrix}
\usage{
npz.to.spM(npzPath)
}
\arguments{
\item{npzPath}{Directory of npz file.}
}
\value{
Tsparse matrix with j = indices, p = indptr and x = data.
}
\description{
\code{npz.to.spM} converts a npz file to a Tsparse matrix.
}
\examples{
file_paths <- load_exdata()
methylpath <- file_paths[[1]]
npz.to.spM(methylpath)
}
