% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npz.to.spM.R
\name{npz.to.spM}
\alias{npz.to.spM}
\title{Transform npz file to sparse Matrix}
\usage{
npz.to.spM(npzPath)
}
\arguments{
\item{npzPath}{pathway of npz file}
}
\value{
Tsparse Matrix with j = indices, p = indptr and x = data
}
\description{
Transform npz file to sparse Matrix
}
