% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{plot_all}
\alias{plot_all}
\title{Plotting of CpG, GpC and gene data}
\usage{
plot_all(
  cpgifile,
  npz,
  meta,
  header,
  npz_acc,
  header_acc,
  featurefile,
  genefile,
  hx,
  ht,
  chr,
  startpos,
  endpos,
  start_VR,
  end_VR
)
}
\arguments{
\item{cpgifile}{path to txt file containing information about CpG islands retrieved from ...}

\item{npz}{pathway of npz file}

\item{meta}{data frame of meta data containing cell IDs ("cell_id_dna") and pseudo time ("ptime)}

\item{header}{data frame containing cell IDs for spM}

\item{npz_acc}{pathway of npz file containing GpC methylation. Used for chromatin accessibility analysis}

\item{header_acc}{data frame containing cell IDs for spM}

\item{featurefile}{pathway of gtf file downloaded from ensemble containing ...}

\item{genefile}{pathway to gtf file downloaded from...}

\item{hx}{width of the kernel}

\item{ht}{height of the kernel}

\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}

\item{start_VR}{integer defining the start position of the variable region}

\item{end_VR}{integer defining the end position of the variable region}
}
\value{
Plot
}
\description{
Plotting of CpG, GpC and gene data
}
