% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{plot_all}
\alias{plot_all}
\title{Create summary plot of CpG and GpC methylation data with genomic information}
\usage{
plot_all(
  species,
  cpgipath,
  spM,
  meta,
  header,
  spMacc,
  header_acc,
  featurepath,
  genepath,
  hx,
  ht,
  chr,
  startpos,
  endpos,
  start_VR,
  end_VR
)
}
\arguments{
\item{species}{String of species name/alias.}

\item{cpgipath}{Character string containing pathway to text file. Alternatively, can be connection. File contains information about CpG islands retrieved from https://genome.ucsc.edu/cgi-bin/hgTables. Use for "table" either cpgIslandExt or cpgIslandExtUnmasked.}

\item{spM}{dgTMatrix containing CpG methylation.}

\item{meta}{data frame of meta data containing cell IDs ("cell_id_dna") and pseudo time ("ptime")}

\item{header}{data frame containing cell IDs for spM}

\item{spMacc}{dgTMatrix containing GpC methylation. Used for chromatin accessibility analysis}

\item{header_acc}{data frame containing cell IDs for spM of accessibility data?}

\item{featurepath}{Location of the gff file to be read. Can be a single string of the file path or the URL or can be a connection.}

\item{genepath}{Location of the gtf file to be read. Can be a single string containing the file path or the URL or can be a connection.}

\item{hx, ht}{Width and height of kernel.}

\item{chr}{Integer number of chromosome}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}

\item{start_VR, end_VR}{Integers defining the start and end position of the variable region.}
}
\value{
Plot
}
\description{
Create summary plot of CpG and GpC methylation data with genomic information
}
\examples{
\dontrun{plot_all("mouse", cpgipath, spM, meta, header, spMacc, header_acc, featurepath, genepath, 400, 0.08, 8, 8628165, 8684055, 8653165, 8659055)}
}
