% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CpGislands.R
\name{plot_CpGislands}
\alias{plot_CpGislands}
\title{Plot CpG islands}
\usage{
plot_CpGislands(path, chr, startpos, endpos)
}
\arguments{
\item{path}{path to txt file containing information about CpG islands retrieved from ...}

\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}
}
\value{
segment plot
}
\description{
Plot CpG islands
}
