% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpgislands.R
\name{plot_cpgislands}
\alias{plot_cpgislands}
\title{Plot CpG islands}
\usage{
plot_cpgislands(genome, chr, startpos, endpos)
}
\arguments{
\item{genome}{String of the genome version used.}

\item{chr}{Integer number of chromosome.}

\item{startpos, endpos}{Integers defining the start and end position of the analysed genomic region.}
}
\value{
Segment plot indicating the position of CpG islands within the genomic region.
}
\description{
\code{plot_cpgislands} downloads and plots CpG islands of the chosen genomic region as a segment plot visualizing the position of islands.
Information of the CpG islands are obtained from the UCSC Genome Browser \url{https://genome.ucsc.edu/cgi-bin/hgTables}.
}
\examples{
plot_cpgislands("mm39", 8, 8628165, 8684055)
}
