% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CpGislands.R
\name{plot_CpGislands}
\alias{plot_CpGislands}
\title{Plot cpg islands}
\usage{
plot_CpGislands(txtpath, chr, startpos, endpos)
}
\arguments{
\item{txtpath}{Character string containing pathway to text file. Alternatively, can be connection. File contains information about CpG islands retrieved from https://genome.ucsc.edu/cgi-bin/hgTables . Use for "table" either cpgIslandExt or cpgIslandExtUnmasked.}

\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}
}
\value{
segment plot
}
\description{
Plot cpg islands
}
\examples{
# dont run
plot_CpGislands("~/gpgi_file", 8, 8628165, 8684055)
}
