% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpgs.R
\name{plot_cpgs}
\alias{plot_cpgs}
\title{Get bar plot visualizing amounts of CpGs within a genomic region}
\usage{
plot_cpgs(chr, startpos, endpos)
}
\arguments{
\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}
}
\value{
bar plot
}
\description{
Get bar plot visualizing amounts of CpGs within a genomic region
}
\examples{
plot_cpgs(8, 8628165, 8684055)
}
