% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CpGs.R
\name{plot_CpGs}
\alias{plot_CpGs}
\title{Get bar plot visualizing amounts of CpGs within a genomic region}
\usage{
plot_CpGs(chr, start, end)
}
\arguments{
\item{chr}{integer number chromosome}

\item{start}{integer defining the start position of the analysed genomic region}

\item{end}{integer defining the end position of the analysed genomic region}
}
\value{
bar plot
}
\description{
Get bar plot visualizing amounts of CpGs within a genomic region
}
