% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genemodel.R
\name{plot_genemodel}
\alias{plot_genemodel}
\title{Plot gene model}
\usage{
plot_genemodel(gtfpath, chr, startpos, endpos)
}
\arguments{
\item{gtfpath}{Location of the gtf file to be read. Can be a single string containing the file path or the URL or can be a connection.}

\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}
}
\value{
arrow plot
}
\description{
Plot gene model
}
\examples{
get_genemodel("https://ftp.ensembl.org/pub/release-110/gtf/mus_musculus/Mus_musculus.GRCm39.110.gtf.gz", 8,  8628165, 8684055)
}
