% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genemodel.R
\name{plot_genemodel}
\alias{plot_genemodel}
\title{Plot gene model}
\usage{
plot_genemodel(gtf, start_pos, end_pos)
}
\arguments{
\item{gtf}{pathway to gtf file downloaded from...}

\item{start_pos}{integer defining the start position of the analysed genomic region}

\item{end_pos}{integer defining the end position of the analysed genomic region}
}
\value{
arrow plot
}
\description{
Plot gene model
}
