% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genemodel.R
\name{plot_genemodel}
\alias{plot_genemodel}
\title{Plot gene model}
\usage{
plot_genemodel(gtf, chr, startpos, endpos)
}
\arguments{
\item{gtf}{pathway to gtf file downloaded from...}

\item{chr}{integer number of chromosome}

\item{startpos}{integer defining the start position of the analysed genomic region}

\item{endpos}{integer defining the end position of the analysed genomic region}
}
\value{
arrow plot
}
\description{
Plot gene model
}
