% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methyl.R
\name{plot_methyl}
\alias{plot_methyl}
\title{Plot mapped methylation data}
\usage{
plot_methyl(m)
}
\arguments{
\item{m}{matrix received from function \emph{map_methyl}}
}
\value{
tile plot
}
\description{
Plot mapped methylation data
}
