% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methyl.R
\name{plot_methyl}
\alias{plot_methyl}
\title{Plot methylation data}
\usage{
plot_methyl(m, mappedpt, startpos)
}
\arguments{
\item{m}{Matrix received from \code{smooth2d}. Contains smoothed values and pseudotime of scNMT-seq data of cells at various differentiation stages.}

\item{mappedpt}{Data frame of the sparse matrix received from \code{map_methyl}. Contains relative genomic position (pos), pseudotime of the cells (pt) and methylation status (methyl) -1 or 1 (un-/methylated) of the selected region, sorted by pseudotime.}

\item{startpos}{Integer defining the start position of the analysed genomic region.}
}
\value{
Tile plot visualizing methylation pattern of a genomic region at various differentiation stages.
}
\description{
\code{plot_methyl} generates a tile plot to visualize the smoothed methylation values from temporal scNMT-seq data, highlighting changes during differentiation.
}
\examples{
\dontrun{plot_methyl(m, mappedpt, 8628165)}
}
