% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methyl.R
\name{plot_methyl}
\alias{plot_methyl}
\title{Plot methylation data}
\usage{
plot_methyl(m, mappedpt, startpos, endpos)
}
\arguments{
\item{m}{Matrix received from \code{smooth}. Contains smoothed values and pseudotime of scNMT-seq data of cells at various differentiation stages.}

\item{mappedpt}{Data frame of the sparse matrix obtained from \code{map_methyl}. Contains relative genomic position (pos), pseudotime of the cells (pt) and methylation status (methyl), which is -1 or 1 (un-/methylated) for the selected region of the scNMT-seq data.}
}
\value{
Tile plot visualizing methylation pattern of a genomic region at various differentiation stages.
}
\description{
\code{plot_methyl} generates a tile plot to visualize the smoothed methylation values from temporal scNMT-seq data, highlighting changes during differentiation.
}
\examples{
# create example data frame
pos = c(5618, 5619, 5620,5621, 5622, 5623, 5624,5625, 5626,5627)
pt = c(1,1,2,2,3,3,4,4,5,5)
methyl = c(1,1,1,-1,1,-1,-1,1,-1,-1)
data.frame( pos, pt, methyl) -> mappedpt

# create matrix
smooth(mappedpt, 2, 2, 0.1, 0.1) -> m

plot_methyl(m, mappedpt, 8628165, 8684055)
}
