% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methyl.R
\name{plot_methyl}
\alias{plot_methyl}
\title{Plot of mapped methylation data}
\usage{
plot_methyl(m, mappedpt, startpos)
}
\arguments{
\item{m}{matrix received from \emph{gauss_kernel_2d}}

\item{mappedpt}{data frame containing pseudo times received from \emph{map_methyl}}

\item{startpos}{integer defining the start position of the analysed genomic region}
}
\value{
tile plot
}
\description{
Plot of mapped methylation data
}
\examples{
 # dont run
plot_methyl(m, mappedpt, 8628165)
}
