% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regfeatures.R
\name{plot_regfeatures}
\alias{plot_regfeatures}
\title{Plot regulatory features of genomic region}
\usage{
plot_regfeatures(
  regfeatures,
  chr,
  start_pos,
  end_pos,
  start_VR = 0,
  end_VR = 0
)
}
\arguments{
\item{regfeatures}{pathway of gtf file downloaded from ensemble containing ...}

\item{start_pos}{integer defining the start position of the analysed genomic region}

\item{end_pos}{integer defining the end position of the analysed genomic region}

\item{start_VR}{integer defining the start position of the variable region}

\item{end_VR}{integer defining the end position of the variable region}
}
\value{
segment plot
}
\description{
Plot regulatory features of genomic region
}
